Allow non-absolute path for variables. This fixes grit to do variable substitution in the resource_ids file correctly when it is invoked by ninja (which uses a relative path for this, whereas our gyp-generated make files use an absolute path).
This mirrors Chrome r102327 to grit-i18n. This was an issue that came
up after I snapshotted grit from Chrome but before I removed the grit
sources on the Chrome side, so the fix needed to be made on the Chrome
side and migrated over.
Review URL: http://codereview.chromium.org/7995016

git-svn-id: http://grit-i18n.googlecode.com/svn/trunk@5 7262f16d-afe8-6277-6482-052fa10e57b1