Allow substitution of messages as variables in other messages.
This is part of a manual merge of changes released to the internal branch 2008/06/11.  Original author is Marc-Antoine Parent.

In addition to the merge, bring in all language directive mappings and codepage mappings from the latest internal version of grit to avoid warnings related to unknown output languages.  While updating the mapping, fix the name of the fake-bidi language.  Finally, to help debug differences in data_pack files, add a simple ASCII converter to data_pack.py

BUG=http://code.google.com/p/grit-i18n/issues/detail?id=4

Review URL: https://chromiumcodereview.appspot.com/9965022

git-svn-id: http://grit-i18n.googlecode.com/svn/trunk@26 7262f16d-afe8-6277-6482-052fa10e57b1