Replace SetOutputContext with SetOutputLanguage in GritNode
I am about to add a new attribute called "context" to output tags, so the old
method name will be confusing. Unlike the language, the defines don't change
within a single run (except in unit tests), so it's not necessary to set them
every time the language is changed. There is already a SetDefines() method.

There appears to have been a bug in the old code: when SetOutputContext() was
called more than once with different defines, the substituter was updated with
the new substitutions instead of being cleared and recreated. The new code
recreates it from scratch.

It might be possible (and better?) to get rid of SetDefines() and always pass
the defines as a parameter on creation, or even place the defines in a global
variable or singleton (like sys.platform, which is used in a similar way in
the code).

BUG=None
TEST=Unit tests, rebuilt Chrome

Review URL: https://chromiumcodereview.appspot.com/10535068

git-svn-id: http://grit-i18n.googlecode.com/svn/trunk@47 7262f16d-afe8-6277-6482-052fa10e57b1