Refactor chrome_html and add conversion of img tags.
Adds a style attribute to <img> tags which don't contain a style (for safety)
giving them available high dpi image assets automatically. We preserve the
src in case this is referenced in script. Additionally this changes inlining
and image-set conversions to preserve the type of quotation used.

BUG=chromium:135179, chromium:135779
TEST=ChromeHtmlUnittest.testFileResourcesImageTag
Review URL: https://chromiumcodereview.appspot.com/10824026


git-svn-id: http://grit-i18n.googlecode.com/svn/trunk@63 7262f16d-afe8-6277-6482-052fa10e57b1