Allow multiple use_name_for_id messages with identical ID.
Normally, messages with different contents get different IDs (in the
Clique or XTB ID sense), but separate messages using use_name_for_id
may have different contents, so the assertion was incorrect.

Also in this change, unify the expression evaluation used for <if>
nodes in .xtb files with that in .grd files. This is unrelated to the
bug referenced below, but I noted that a pending patch wants to use
some of the variables previously unavailable to expressions in .xtb
files, and it makes sense to unify it. Output context and language do
not make sense when evaluating an .xtb file and hence are still only
available to expressions in .grd files.

BUG=https://code.google.com/p/grit-i18n/issues/detail?id=8

Review URL: https://codereview.chromium.org/12036015

git-svn-id: http://grit-i18n.googlecode.com/svn/trunk@105 7262f16d-afe8-6277-6482-052fa10e57b1