Add the ability to generate a depfile when running grit build.
This is analogous to the ability of some compilers to generate depfiles
that contain the .h files a .cc file depends on when it compiles it.
This is useful for speeding up build systems because the input
dependencies don't need to be determined before compilation.

As there are no pre-determined outputs from a .grd file (all the outputs
are determined by the file contents) the depfile refers to itself and
acts as a stamp for the generated outputs.

BUG=297672

Original author koz@chromium.org. Original patch is issue 25683011.

Review URL: https://codereview.chromium.org/27001002

git-svn-id: http://grit-i18n.googlecode.com/svn/trunk@137 7262f16d-afe8-6277-6482-052fa10e57b1